transforms-test success was a false positive but made some changes
The test for valid input was passing because we just called the function without an expect statement. I added a couple lines to the randomColors function to throw an error if it detects edge cases. The test for invalid inputs works, but checking to see if the color array changes is a roadblock, at least for just that test. We should still make more transform functions like grey-scale or color-tint and we'll ask about the roadblock on Monday. I suggest we make tests for the other transforms in their own modules because they are long. I also renamed the two test-output image names so they'll make more sense. I suggest we do any other transforms we want while the input/output files are hardcoded in index.js, and leave argv for last (which we DON'T need a test for).